fix(quill): clean up charts package public API and core layering
- drop computeBoxBand/computeBoxRect/computeSeriesBoxes from the barrel
  (internal helpers typed with d3-scale params, zero external consumers)
- move the public box-plot types (BoxPlotDatum/BoxPlotSeries to a d3-free
  BoxPlot/types.ts, BoxRect to core/types.ts) so the published d.ts no
  longer reaches a d3-scale import; @types/d3-scale stays a devDependency
- move PieLayout/PieSlice/sliceAt/cursorOffsetToAngle to core/radial-layout
  so core/ no longer imports from charts/ (re-exported for back-compat)

Generated-By: PostHog Code
Task-Id: 81cfc552-8a48-447d-8332-3a643651f29c
