+++ packages/quill/packages/charts/src/charts/BoxPlot/BoxPlot.tsx
@@ -18,7 +18,8 @@ import type {
     TooltipContext,
 } from '../../core/types'
 import { BoxPlotTooltip } from './BoxPlotTooltip'
-import { computeBoxRect, computeSeriesBoxes, type BoxPlotDatum, type BoxPlotSeries } from './computeBoxLayout'
+import { computeBoxRect, computeSeriesBoxes } from './computeBoxLayout'
+import type { BoxPlotDatum, BoxPlotSeries } from './types'
 import { seriesKeysAtCursor } from './utils/boxes-under-cursor'
 
 /** Stash slot — survives a render via `ChartScales._private` so drawStatic/drawHover
